netology_pd_diplom/requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.0.6
django-rest-passwordreset==1.4.1
djangorestframework==3.14.0
gunicorn==22.0.0
idna==3.7
iniconfig==2.0.0
kombu==5.3.7
model-bakery==1.18.0
packaging==24.0
pluggy==1.5.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pytest==8.2.1
pytest-django==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==4.5.1
requests==2.31.0
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
ujson==5.9.0
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
